Rename authController to clientsController. refactore xml-doc in clientsController
